Fixed a bug in the client request/renew cert logic
If the client tried to renew its certificate but was rejected by
the server, it would give up.

In a situation where the certificate is actually invalid, it will
never be accepted by the server, and the situation would not resolve
itself.

Now it tries first to renew, and if that fails it tries to request
a new certificate. The old certificate will be kept until a new one
is successfully received.