Work in progress on the new in-memory search
processarchive creates tasks for removing obsolete files from the
search cache.
Taskrunner has been reverted back to regular http, to keep it simple.
There's a new job for verifying that the cache is up-to-date.
Bugfix in the createExcerpt-function in fastSearch.
Minor improvements to the search page.
This commit is work in progress on #51.