New API and button on web to retry failed tasks
Reason: Failed tasks are retried periodically.
If the reason for failure has been fixed,
you may want to retry tasks sooner than the system otherwise would.

The method works by settings the delay fields to 0
in the database table, so the tasks will be retried as soon
as the task runner reloads the table.

The API method requires admin rights.