Hardcode the version number in the client
The client used to load its version number from /etc/nivlheim/version.
This created a potential problem if a package upgrade failed to
update the version file, or it got changed by another process.
It also created a problem on the server, where 2 packages claimed
to own the same file.

It's better if the version number is hardcoded in the client script,
that ensures that the client always reports the correct version.