Bugfix: hostlist API internal error
This commit fixes a bug where the hostlist API would make an internal
error if you tried to update custom fields for nonexistent hosts.
Fixes #114.

Also, the hostlist API will now continue to process hosts even if
there are errors with some hosts. In that case it will return a list
of error messages, and http status 200 OK.