Add random sleep to Windows client, and fix FQDN
This commit adds a random sleep interval between 0 and 55 minutes
at the start, to avoid a situation where all the clients try to
contact the server at the same time every hour.

Also, the code that determines the hostname is changed because we want
the FQDN, not just the hostname. This commit fixes #138.