Removal of duplicates will not consider IP address
Nivlheim can end up with duplicates of a particular machine if it is
re-installed and the certificate is lost.
The code that detects these duplicates look at hardware information
(serial number, product) as well as OS hostname and IP address.

We've observed a few cases where the IP address was changed during
reinstallation. Therefore it makes sense to remove that condition.