Bugfix: renewcert must notify the system service
This commit fixes a bug where renewcert would give a machine a new
certificate and update all its files in the database, but the memory
cache in the system service would still have the files associated with
the old certificate. This resulted in the API function "grep" returning
results with an empty string instead of the hostname.