Overwrite cache entries instead of remove + add
When files change, the ram cache must be updated. This commit changes
the way it is done from remove, then add, to just overwrite.
This ensures that the cache contains the file at all times. Fixes #144.

Also, the compareSearchCacheToDatabase function is re-written to be more
resilient to concurrent changes by other parts of the software.