Fix the way the utility package is imported
Discovered that Go used a separate, downloaded copy of the utility
package when compiling, instead of using local files.
What we want is for it to use the files in the utility subdirectory.
Turns out, we (I) had been importing the utility package the wrong way
this whole time.
This commit fixes go.mod and every Go file that imports the package.