Change nivlheim_client regex parse in split_url
Nivlheim_client has a function which splits the server URL in several variables that contain the protocol, URL, port and path. If you passed a complex URL with subdomains (like example01.test.uib.cat) or without domain name the function was not able to parse them.

Replaced the regex to allow more complex domain names and changed the regex groups that are assigned to variables. The trivial case with localhost fits in this regex, so the last "if" is not needed anymore. Checked with localhost and other domains and seems to be working.