Bugfix: Files are wrongly removed from search
Some code in processarchive.go was placed in the processFile function,
this was probably by mistake, it should go in the processArchive
function instead.

The result was that when processing an archive, all of the host's
current files were removed from search cache. They would be added back
by the compareSearchCacheToDB function a few minutes later.